Streamlit/requirements.txt
psycopg2-binary
streamlit
pandas
bcrypt
openai
pinecode